feat: implement character crud with skills management
- Add character creation, read, update, delete operations
- Implement character skills association (add, list, remove)
- Fix CharacterAttribute cascade issues by removing trigger
- Organize CharacterController with proper route precedence
- Add CharacterSkillId as standalone class for composite keys
- All endpoints tested and working

+++ src/main/java/com/mugen/backend/dto/CharacterDTO.java
@@ -24,5 +24,5 @@ public class CharacterDTO {
     private String name;
 
     @NotNull(message = "Race ID is required")
-    private Long raceId;
+    private Integer raceId;
 }
